# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  db.CurrencyRate:
    properties:
      buy_rate:
        type: number
      currency:
        type: string
      sell_rate:
        type: number
      standard_rate:
        type: number
      updated_at:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            example: VALIDATION_ERROR
            type: string
          message:
            example: phone is required
            type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        example: OTP sent
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  handlers.PaymentsWebhookData:
    properties:
      address:
        type: string
      amount:
        type: string
      confirmations:
        type: integer
      confirmed_at:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      detected_at:
        type: string
      network:
        type: string
      safe_to_credit:
        type: boolean
      status:
        type: string
      tx_hash:
        type: string
    type: object
  handlers.PaymentsWebhookPayload:
    properties:
      data:
        $ref: '#/definitions/handlers.PaymentsWebhookData'
      event:
        type: string
    type: object
  handlers.ProfileData:
    properties:
      avatar_url:
        type: string
      bio:
        example: Digital finance enthusiast from Cameroon
        type: string
      bvn:
        example: "12345678901"
        type: string
      date_of_birth:
        example: "1995-06-15"
        type: string
      email:
        example: alice@example.com
        type: string
      first_name:
        example: Alice
        type: string
      id:
        type: string
      is_email_verified:
        example: false
        type: boolean
      kyc_status:
        example: pending
        type: string
      last_name:
        example: Dupont
        type: string
      nationality:
        example: Cameroonian
        type: string
      phone_number:
        example: "+237612345678"
        type: string
    type: object
  handlers.ProfileResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.ProfileData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.SetCurrencyRateRequest:
    properties:
      buy_rate:
        example: 615
        type: number
      sell_rate:
        example: 603
        type: number
      standard_rate:
        example: 609
        type: number
    type: object
...
  handlers.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        example: Digital finance enthusiast
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rach.finance
    name: Rach Finance Engineering
  description: |-
    DigitalFX is a multi-currency fintech platform for Central/West Africa.
    It combines fiat accounts (XAF, XOF, USD, GBP, EUR), HD crypto wallets (BIP-44 via Rach WaaS),
    and stablecoin P2P transfers (USDC/USDT via Rach CaaS ERC-4337 SCWs).
    Mobile Money deposits and withdrawals are handled via HUB2 (MTN, Orange).
  title: DigitalFX API
  version: "1.0"
paths:
...
  /admin/currency-rates:
    get:
      description: Admin view of all currency buy/sell/standard rates. Requires the
        rates:view permission.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CurrencyRate'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List currency rates (admin)
      tags:
      - admin
  /admin/currency-rates/{currency}:
    put:
      consumes:
      - application/json
      description: Creates or updates the standard, buy and sell rate for a currency.
        Buy low / sell high for margin; standard is the display rate. Requires the
        rates:manage permission.
      parameters:
      - description: Currency code (USD, EUR, GBP, XAF, XOF, NGN)
        in: path
        name: currency
        required: true
        type: string
      - description: Standard / buy / sell rates (per 1 USD)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SetCurrencyRateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.CurrencyRate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set a currency's rates (admin)
      tags:
      - admin
...
  /profile:
    get:
      description: Returns the full profile of the authenticated user including bio,
        avatar, and KYC/email verification status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: 'Updates the authenticated user''s editable profile fields: bio
        and avatar URL ONLY. Name, phone, email and KYC-verified identity fields (date
        of birth, nationality, BVN) are LOCKED — they cannot be changed here and are
        ignored if sent. Identity corrections go through KYC / admin review.'
      parameters:
      - description: Editable profile fields (bio, avatar_url)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - profile
...
  /rates:
    get:
      description: Returns the buy, sell and standard rate for every supported currency
        (EUR, GBP, USD, XAF, XOF, NGN). Rates are units of that currency per 1 USD.
        Use `standard_rate` to display a customer's balance in another currency; use
        `buy_rate` / `sell_rate` when the customer sells to / buys from the platform.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CurrencyRate'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List currency rates
      tags:
      - rates
  /rates/{currency}:
    get:
      description: Returns the buy, sell and standard rate for a single currency.
      parameters:
      - description: Currency code (USD, EUR, GBP, XAF, XOF, NGN)
        in: path
        name: currency
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.CurrencyRate'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a currency's rates
      tags:
      - rates
...
  /webhooks/payments:
    post:
      consumes:
      - application/json
      description: 'Receiver for Rach WaaS on-chain deposit events. Fires `wallet.deposit.detected`
        (seen on-chain, not credited) and `wallet.deposit.confirmed` (block-depth/settlement
        threshold met). On a confirmed event the owning user (resolved by derived
        address) is notified and the deposit is mirrored best-effort into their local
        ledger; the WaaS balance API remains authoritative. Signature: header `X-Webhook-Signature`
        = hex HMAC-SHA256 of the raw request body, keyed with the WaaS webhook secret
        (PAYMENTS_WEBHOOK_SECRET). Always returns 200 on business-logic issues to
        avoid provider retries.'
      parameters:
      - description: hex HMAC-SHA256(rawBody, secret)
        in: header
        name: X-Webhook-Signature
        required: true
        type: string
      - description: Event name (e.g. wallet.deposit)
        in: header
        name: X-Webhook-Event
        type: string
      - description: Webhook payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.PaymentsWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Rach WaaS Webhook (crypto deposits)
      tags:
      - webhooks
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token — prefix with "Bearer ": Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"